requirements.txt
streamlit
python-dotenv
langchain
langchain_google_genai
pandas